third_party/WebKit/LayoutTests/svg/dom/SVGLengthList-removeItem.html

<!DOCTYPE HTML>
<title>SVGLengthList, removeItem()</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg width="200" height="200">
    <text x="500 50 100 900 150" y="50">ABC</text>
</svg>
<script>
test(function() {
  // This is a test of the SVGLengthList.removeItem() API.
  var list = document.querySelector("text").x.baseVal;

  // Check initial list state of text.
  assert_equals(list.numberOfItems, 5);
  assert_equals(list.getItem(0).value, 500);
  assert_equals(list.getItem(1).value, 50);
  assert_equals(list.getItem(2).value, 100);
  assert_equals(list.getItem(3).value, 900);
  assert_equals(list.getItem(4).value, 150);
  assert_throws("IndexSizeError", function() { list.getItem(5); });

  // Remove first item.
  assert_equals(list.removeItem(0).value, 500);

  // Remove third item.
  assert_equals(list.removeItem(2).value, 900);
});
</script>
